backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
twilio==9.3.0
apscheduler==3.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

